+++ eras/shelley/test-suite/bench/BenchValidation.hs
@@ -5,10 +5,11 @@
 {-# LANGUAGE GADTs #-}
 {-# LANGUAGE RankNTypes #-}
 {-# LANGUAGE ScopedTypeVariables #-}
+{-# LANGUAGE TypeApplications #-}
 {-# LANGUAGE TypeFamilies #-}
 {-# LANGUAGE TypeOperators #-}
+{-# LANGUAGE UndecidableInstances #-}
 {-# OPTIONS_GHC -Wno-orphans #-}
-{-# OPTIONS_GHC -Wno-unused-binds #-}
 
 module BenchValidation (
   ValidateInput (..),
@@ -17,13 +18,14 @@ module BenchValidation (
   benchreValidate,
   applyBlock,
   sizes,
+  UpdateInputs,
   updateChain,
   updateAndTickChain,
   genUpdateInputs,
 ) where
 
 import Cardano.Ledger.BaseTypes (Globals (..), unBlocksMade)
-import Cardano.Ledger.Block (Block (..))
+import Cardano.Ledger.Block (Block (..), EraBlockHeader)
 import qualified Cardano.Ledger.Shelley.API as API
 import Cardano.Ledger.Shelley.Bench.Gen (genBlock, genChainState)
 import Cardano.Ledger.Shelley.BlockBody (slotToNonce)
@@ -38,13 +40,15 @@ import Cardano.Protocol.Crypto
 import Cardano.Protocol.TPraos.API (
   ChainDepState (..),
   ChainTransitionError,
-  GetLedgerView,
-  LedgerView (..),
-  currentLedgerView,
+  TPraosLedgerView,
+  forecastToTPraosLedgerView,
   tickChainDepState,
   updateChainDepState,
  )
-import Cardano.Protocol.TPraos.BHeader (BHeader (..), LastAppliedBlock (..), makeHeaderView)
+import Cardano.Protocol.TPraos.BHeader (
+  BHeader (..),
+  LastAppliedBlock (..),
+ )
 import Cardano.Protocol.TPraos.Rules.Prtcl (PrtclState (..))
 import Cardano.Protocol.TPraos.Rules.Tickn (TicknState (..))
 import Cardano.Slotting.Slot (withOriginToMaybe)
@@ -53,6 +57,7 @@ import Control.Monad.Except ()
 import Control.State.Transition
 import qualified Data.Map.Strict as Map
 import Data.Proxy
+import Test.Cardano.Ledger.BlockHeader (TestBlockHeader (..), mkTestBlockHeaderNoNonce)
 import Test.Cardano.Ledger.Shelley.ConcreteCryptoTypes (MockCrypto)
 import Test.Cardano.Ledger.Shelley.Constants (defaultConstants)
 import Test.Cardano.Ledger.Shelley.Generator.Core (GenEnv)
@@ -76,9 +81,10 @@ validateInput ::
   , EraStake era
   , EraRule "LEDGERS" era ~ API.ShelleyLEDGERS era
   , QC.HasTrace (API.ShelleyLEDGERS era) (GenEnv MockCrypto era)
-  , API.ApplyBlock era
-  , GetLedgerView era
+  , API.ApplyBlock TestBlockHeader era
+  , API.ShelleyEraForecast era
   , MinLEDGER_STS era
+  , EraBlockHeader (BHeader MockCrypto) era
   ) =>
   Int ->
   IO (ValidateInput era)
@@ -89,55 +95,52 @@ validateInput n = do
   pure (ValidateInput testGlobals (chainNes chainstate) block)
 
 benchValidate ::
-  (API.ApplyBlock era, Show (PredicateFailure (EraRule "BBODY" era))) =>
+  ( API.ApplyBlock TestBlockHeader era
+  , Show (PredicateFailure (EraRule "BBODY" era))
+  , EraBlockHeader (BHeader MockCrypto) era
+  ) =>
   ValidateInput era ->
   IO (NewEpochState era)
-benchValidate (ValidateInput globals state (Block bh txs)) =
-  let block = Block (makeHeaderView bh Nothing) txs
-   in case API.applyBlockEitherNoEvents ValidateAll globals state block of
-        Right x -> pure x
-        Left x -> error (show x)
+benchValidate (ValidateInput globals state blk@(Block _ txs)) =
+  case API.applyBlockEitherNoEvents ValidateAll globals state (Block (mkTestBlockHeaderNoNonce blk) txs) of
+    Right x -> pure x
+    Left x -> error (show x)
 
 applyBlock ::
   forall era.
-  ( API.ApplyBlock era
+  ( API.ApplyBlock TestBlockHeader era
   , NFData (StashedAVVMAddresses era)
   , NFData (InstantStake era)
   , GovState era ~ ShelleyGovState era
   , EraCertState era
   , Show (PredicateFailure (EraRule "BBODY" era))
+  , EraBlockHeader (BHeader MockCrypto) era
   ) =>
   ValidateInput era ->
   Int ->
   Int
-applyBlock (ValidateInput globals state (Block bh txs)) n =
-  let block = Block (makeHeaderView bh Nothing) txs
-   in case API.applyBlockEitherNoEvents ValidateAll globals state block of
-        Right x -> seq (rnf x) (n + 1)
-        Left x -> error (show x)
+applyBlock (ValidateInput globals state blk@(Block _ txs)) n =
+  case API.applyBlockEitherNoEvents ValidateAll globals state (Block (mkTestBlockHeaderNoNonce blk) txs) of
+    Right x -> seq (rnf x) (n + 1)
+    Left x -> error (show x)
 
 benchreValidate ::
-  API.ApplyBlock era =>
+  (API.ApplyBlock TestBlockHeader era, EraBlockHeader (BHeader MockCrypto) era) =>
   ValidateInput era ->
   NewEpochState era
-benchreValidate (ValidateInput globals state (Block bh txs)) =
-  API.applyBlockNoValidaton globals state (Block (makeHeaderView bh Nothing) txs)
-
--- ==============================================================
+benchreValidate (ValidateInput globals state blk@(Block _ txs)) =
+  API.applyBlockNoValidaton globals state (Block (mkTestBlockHeaderNoNonce blk) txs)
 
 data UpdateInputs
   = UpdateInputs
       !Globals
-      !LedgerView
+      !TPraosLedgerView
       !(BHeader MockCrypto)
       !ChainDepState
 
 instance Show UpdateInputs where
   show (UpdateInputs _globals vl bh st) =
     show vl ++ "\n" ++ show bh ++ "\n" ++ show st
-
-instance NFData LedgerView where
-  rnf (LedgerView _D _extraEntropy _pool _delegs _ccd) = ()
 
 -- TODO: move upstream
 instance Crypto c => NFData (BHeader c) where
@@ -158,10 +161,11 @@ genUpdateInputs ::
   ( EraGen era
   , EraStake era
   , MinLEDGER_STS era
-  , GetLedgerView era
+  , API.ShelleyEraForecast era
   , EraRule "LEDGERS" era ~ API.ShelleyLEDGERS era
   , QC.HasTrace (API.ShelleyLEDGERS era) (GenEnv MockCrypto era)
-  , API.ApplyBlock era
+  , API.ApplyBlock TestBlockHeader era
+  , EraBlockHeader (BHeader MockCrypto) era
   ) =>
   Int ->
   IO UpdateInputs
@@ -169,7 +173,7 @@ genUpdateInputs utxoSize = do
   let ge = genEnv (Proxy :: Proxy era) defaultConstants
   chainstate <- genChainState utxoSize ge
   (Block blockheader _) <- genBlock ge chainstate
-  let ledgerview = currentLedgerView (chainNes chainstate)
+  let forecast = API.currentForecast (chainNes chainstate)
   let (ChainState _newepochState keys eta0 etaV etaC etaH slot) = chainstate
   let prtclState = PrtclState keys eta0 etaV
   let ticknState = TicknState etaC etaH
@@ -179,7 +183,7 @@ genUpdateInputs utxoSize = do
   pure
     ( UpdateInputs
         testGlobals
-        ledgerview
+        (forecastToTPraosLedgerView forecast)
         blockheader
         (ChainDepState prtclState ticknState nonce)
     )
@@ -193,6 +197,5 @@ updateAndTickChain ::
   UpdateInputs ->
   Either (ChainTransitionError MockCrypto) ChainDepState
 updateAndTickChain (UpdateInputs gl lv bh st) =
-  updateChainDepState gl lv bh
-    . tickChainDepState gl lv True
-    $ st
+  updateChainDepState gl lv bh $
+    tickChainDepState gl lv True st