+++ eras/shelley/test-suite/src/Test/Cardano/Ledger/Shelley/Generator/TxAuxData.hs
@@ -1,4 +1,5 @@
 {-# LANGUAGE NamedFieldPuns #-}
+{-# LANGUAGE TypeApplications #-}
 
 module Test.Cardano.Ledger.Shelley.Generator.TxAuxData (
   genMetadata,
@@ -14,11 +15,11 @@ import Cardano.Ledger.Shelley.TxAuxData (
   ShelleyTxAuxData (..),
  )
 import Control.Exception (assert)
-import qualified Data.ByteString.Char8 as BS (length, pack)
 import qualified Data.Map.Strict as Map
 import qualified Data.Text as T (pack)
-import qualified Data.Text.Encoding as T
+import qualified Data.Text.Foreign as T (lengthWord8)
 import Data.Word (Word64)
+import Test.Cardano.Base.Bytes (genByteArray)
 import Test.Cardano.Ledger.Shelley.Constants (Constants (..))
 import Test.QuickCheck (Gen)
 import qualified Test.QuickCheck as QC
@@ -79,7 +80,7 @@ genDatumString =
     n <- QC.choose (0, min sz 64)
     cs <- genUtf8StringOfSize n
     let s = T.pack cs
-    assert (BS.length (T.encodeUtf8 s) == n) $
+    assert (T.lengthWord8 s == n) $
       return (S s)
 
 -- | Produce an arbitrary Unicode string such that it's UTF8 encoding size in
@@ -105,7 +106,7 @@ genDatumBytestring :: Gen Metadatum
 genDatumBytestring =
   QC.sized $ \sz -> do
     n <- QC.choose (0, min sz 64)
-    B . BS.pack <$> QC.vectorOf n QC.arbitrary
+    B <$> genByteArray n
 
 -- | Generate a 'MD.List [Metadatum]'
 --