+++ eras/shelley/test-suite/src/Test/Cardano/Ledger/Shelley/PropertyTests.hs
@@ -12,11 +12,10 @@ module Test.Cardano.Ledger.Shelley.PropertyTests (
   commonTests,
 ) where
 
-import Cardano.Ledger.BHeaderView (BHeaderView)
 import Cardano.Ledger.BaseTypes (Globals, ShelleyBase, SlotNo)
-import Cardano.Ledger.Block (Block)
+import Cardano.Ledger.Block (BbodySignal)
 import Cardano.Ledger.Core
-import Cardano.Ledger.Shelley.API (ApplyBlock, ShelleyPOOL)
+import Cardano.Ledger.Shelley.API (ApplyBlock, ShelleyEraForecast, ShelleyPOOL)
 import Cardano.Ledger.Shelley.Core
 import Cardano.Ledger.Shelley.LedgerState (LedgerState, NewEpochState)
 import Cardano.Ledger.Shelley.Rules (
@@ -28,10 +27,10 @@ import Cardano.Ledger.Shelley.Rules (
   ShelleyPoolPredFailure,
  )
 import Cardano.Ledger.Shelley.State
-import Cardano.Protocol.TPraos.API (GetLedgerView)
 import Cardano.Protocol.TPraos.Rules.Tickn (TicknEnv, TicknState)
 import Control.State.Transition
 import Data.Sequence (Seq)
+import Test.Cardano.Ledger.BlockHeader (TestBlockHeader)
 import qualified Test.Cardano.Ledger.Shelley.ByronTranslation as ByronTranslation (
   testGroupByronTranslation,
  )
@@ -66,8 +65,8 @@ commonTests ::
   ( EraGen era
   , EraStake era
   , ShelleyEraAccounts era
-  , ApplyBlock era
-  , GetLedgerView era
+  , ApplyBlock TestBlockHeader era
+  , ShelleyEraForecast era
   , Embed (EraRule "BBODY" era) (CHAIN era)
   , Embed (EraRule "TICK" era) (CHAIN era)
   , Embed (EraRule "TICKN" era) (CHAIN era)
@@ -78,7 +77,6 @@ commonTests ::
   , Signal (EraRule "LEDGERS" era) ~ Seq (Tx TopTx era)
   , Signal (EraRule "TICKN" era) ~ Bool
   , BaseM (EraRule "LEDGERS" era) ~ ShelleyBase
-  , AtMostEra "Alonzo" era
   , GovState era ~ ShelleyGovState era
   , InstantStake era ~ ShelleyInstantStake era
   , QC.BaseEnv (EraRule "LEDGER" era) ~ Globals
@@ -93,7 +91,7 @@ commonTests ::
   , State (EraRule "LEDGERS" era) ~ LedgerState era
   , Environment (EraRule "BBODY" era) ~ BbodyEnv era
   , Signal (EraRule "TICK" era) ~ SlotNo
-  , Signal (EraRule "BBODY" era) ~ Block BHeaderView era
+  , Signal (EraRule "BBODY" era) ~ BbodySignal era
   , EraRule "POOL" era ~ ShelleyPOOL era
   , InjectRuleFailure "POOL" ShelleyPoolPredFailure era
   , InjectRuleEvent "POOL" PoolEvent era